evals/harbor: collapse into single cmd.py, drop profile, env-based secrets
Restructure the harbor benchmark tooling into a single uv-script entry
point with subcommands run/list/show/task/compare.

Removes:
- goose_harbor/ package layout (runner.py + goose_binary.py)
- tests/ (the tests covered scaffolding that no longer exists)
- pyproject.toml, run shim, .pytest_cache, .venv, .runs/, report.py

Adds:
- cmd.py: PEP 723 inline-uv script, GooseBinaryAgent inlined, .env loader
- config_template.yaml: extension vocabulary; --extensions toggles enabled
- README.md: rewritten for the new layout

Behaviour changes:
- Secrets come from a .env file or the host env, forwarded into the task
  container via harbor's environment.env (same mechanism harbor's pi agent
  uses). The --goose-profile flag is gone; config.yaml is generated inline
  from config_template.yaml.
- New --extensions flag (default: developer,todo) flips enabled flags in
  the template. Unknown extensions error out.
- --tasks replaces repeatable --task, comma-separated.
- Defaults tightened for the common case: dataset terminal-bench-2,
  model anthropic/claude-sonnet-4-6, concurrency 4, max-turns 100,
  install-goose-runtime-deps true.
- Runs land in evals/harbor/runs/ (gitignored), not .runs/jobs/.
