--- evals/harbor/goose_harbor/__init__.py
@@ -1,1 +1,0 @@
-
