--- evals/harbor/pyproject.toml
@@ -1,13 +1,0 @@
-[project]
-name = "goose-harbor-eval"
-version = "0.1.0"
-description = "Goose eval tooling for Harbor benchmark datasets"
-requires-python = ">=3.12"
-dependencies = [
-    "harbor==0.6.4",
-]
-
-[dependency-groups]
-dev = [
-    "pytest>=8.4.0",
-]
