--- evals/harbor/run
@@ -1,12 +1,0 @@
-#!/usr/bin/env sh
-set -eu
-
-SCRIPT_DIR=$(CDPATH= cd -- "$(dirname -- "$0")" && pwd)
-
-if [ -n "${PYTHONPATH:-}" ]; then
-  export PYTHONPATH="$SCRIPT_DIR:$PYTHONPATH"
-else
-  export PYTHONPATH="$SCRIPT_DIR"
-fi
-
-exec python3 "$SCRIPT_DIR/goose_harbor/runner.py" "$@"
